CAP 6 EX 4
=> Add main function for INNER_MODE

=> Alter all arithmetic functions to work with float values

+++ Miguel/Exercicios/Capitulo_6/ex4_operacoes_aritmeticas_basicas/middleware.h
@@ -3,12 +3,12 @@
 
 void clear_buffer();
 
-void ask_for_two_numbers(int *num_1, int *num_2);
+void ask_for_two_numbers(float *num_1, float *num_2);
 
-void print_add_operation(int num_1, int num_2);
+void print_add_operation(float num_1, float num_2);
 
-void print_sub_operation(int num_1, int num_2);
+void print_sub_operation(float num_1, float num_2);
 
-void print_mult_operation(int num_1, int num_2);
+void print_mult_operation(float num_1, float num_2);
 
-void print_div_operation(int num_1, int num_2);
+void print_div_operation(float num_1, float num_2);
